# internal/app/openapi.yaml
openapi: 3.1.0
info:
  title: Kore API
  version: 1.0.0
  description: API Kore — monolithe modulaire PSA/ESN
servers:
  - url: http://localhost:8080/api/v1
paths:
...
  /tickets:
    get:
      summary: Liste tickets support
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: OK }
    post:
      summary: Créer ticket
      security: [{ bearerAuth: [] }]
      responses:
        "201": { description: Créé }
  /tickets/{id}:
    get:
      summary: Détail ticket
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200": { description: OK }
  /tickets/{id}/take-over:
    post:
      summary: Prendre en charge ticket
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200": { description: OK }
  /tickets/{id}/replies:
    post:
      summary: Répondre au ticket
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200": { description: OK }
  /tickets/{id}/resolve:
    post:
      summary: Résoudre ticket
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200": { description: OK }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT